+++ src/main/java/org/betonquest/betonquest/quest/condition/entity/EntityCondition.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.quest.condition.entity;
 
+import net.kyori.adventure.text.Component;
 import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.condition.nullable.NullableCondition;
@@ -37,7 +38,7 @@ public class EntityCondition implements NullableCondition {
      * The name of the entity to check for.
      */
     @Nullable
-    private final Variable<String> name;
+    private final Variable<Component> name;
 
     /**
      * The marked entity to check for.
@@ -55,7 +56,7 @@ public class EntityCondition implements NullableCondition {
      * @param marked        the marked entity to check for
      */
     public EntityCondition(final Variable<List<Map.Entry<EntityType, Integer>>> entityAmounts, final Variable<Location> loc,
-                           final Variable<Number> range, @Nullable final Variable<String> name, @Nullable final Variable<String> marked) {
+                           final Variable<Number> range, @Nullable final Variable<Component> name, @Nullable final Variable<String> marked) {
         this.entityAmounts = entityAmounts;
         this.loc = loc;
         this.range = range;
@@ -67,9 +68,9 @@ public class EntityCondition implements NullableCondition {
     public boolean check(@Nullable final Profile profile) throws QuestException {
         final Location location = loc.getValue(profile);
         final double resolvedRange = range.getValue(profile).doubleValue();
-        final String resolvedName = name == null ? null : name.getValue(profile);
+        final Component name = this.name == null ? null : this.name.getValue(profile);
         final String resolvedMarked = marked == null ? null : marked.getValue(profile);
-        final List<Entity> selectedEntity = EntityUtils.getSelectedEntity(location, resolvedName, resolvedMarked, resolvedRange);
+        final List<Entity> selectedEntity = EntityUtils.getSelectedEntity(location, name, resolvedMarked, resolvedRange);
         for (final Map.Entry<EntityType, Integer> entry : entityAmounts.getValue(profile)) {
             final long count = selectedEntity.stream().filter(entity -> entity.getType() == entry.getKey()).count();
             if (count < entry.getValue()) {