+++ src/main/java/org/betonquest/betonquest/quest/event/spawn/SpawnMobEventFactory.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.quest.event.spawn;
 
+import net.kyori.adventure.text.Component;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.api.quest.event.PlayerEvent;
 import org.betonquest.betonquest.api.quest.event.PlayerEventFactory;
@@ -15,7 +16,6 @@ import org.betonquest.betonquest.instruction.variable.Variable;
 import org.betonquest.betonquest.quest.PrimaryServerThreadData;
 import org.betonquest.betonquest.quest.event.PrimaryServerThreadEvent;
 import org.betonquest.betonquest.quest.event.PrimaryServerThreadPlayerlessEvent;
-import org.betonquest.betonquest.util.Utils;
 import org.bukkit.Location;
 import org.bukkit.entity.EntityType;
 import org.bukkit.entity.Mob;
@@ -61,7 +61,7 @@ public class SpawnMobEventFactory implements PlayerEventFactory, PlayerlessEvent
         final Variable<Location> loc = instruction.get(Argument.LOCATION);
         final Variable<EntityType> type = instruction.get(new EntityTypeParser());
         final Variable<Number> amount = instruction.get(Argument.NUMBER);
-        final Variable<String> name = instruction.getValue("name", value -> Utils.format(value, true, false));
+        final Variable<Component> name = instruction.getValue("name", Argument.MESSAGE);
         final Variable<String> marked = instruction.getValue("marked", PackageArgument.IDENTIFIER);
         final Variable<Item> helmet = instruction.getValue("h", PackageArgument.ITEM);
         final Variable<Item> chestplate = instruction.getValue("c", PackageArgument.ITEM);