+++ src/main/java/org/betonquest/betonquest/util/EntityUtils.java
@@ -1,5 +1,6 @@
 package org.betonquest.betonquest.util;
 
+import net.kyori.adventure.text.Component;
 import org.betonquest.betonquest.BetonQuest;
 import org.bukkit.Location;
 import org.bukkit.NamespacedKey;
@@ -27,14 +28,14 @@ public final class EntityUtils {
      * @param range    the range around the location
      * @return the selected entity
      */
-    public static List<Entity> getSelectedEntity(final Location location, @Nullable final String name,
+    public static List<Entity> getSelectedEntity(final Location location, @Nullable final Component name,
                                                  @Nullable final String marked, final double range) {
         final Collection<Entity> entities = location.getNearbyEntities(range, range, range);
         return entities.stream().filter(entity -> isSelectedEntity(entity, name, marked)).toList();
     }
 
-    private static boolean isSelectedEntity(final Entity entity, @Nullable final String name, @Nullable final String mark) {
-        if (name != null && !entity.getName().equals(name)) {
+    private static boolean isSelectedEntity(final Entity entity, @Nullable final Component name, @Nullable final String mark) {
+        if (name != null && !entity.name().equals(name)) {
             return false;
         }
         if (mark != null) {